Don't crash if unparseable values exist on the failure queue.
We saw some jobs become unloadable on our resque queue, due to issues turning strangely-encoded strings into JSON and back. This meant that resque couldn't parse the jobs, and so we were seeing `nil` elements in the `failures` array.

This is the simplest way of preventing these malformed entries from preventing monitoring of the failures queue entirely.