Fixing/updating the hoptoad backend
The previous version of the hoptoad back end no longer worked. This is due to Hoptoad changing their API somewhere along the way.
I've updated the code here to simply use the hoptoad_notifier gem that folks are likely already using in their apps. This puts the
maintenance and configuration responsibility into that gem rather than here in resque.