Rename Resque::Logger to Resque::Logging
Having both Resque.logger (which expects and stdlib Logger duck type)
and Resque::Logger is confusing.

Resque::Logging is a module that can be added to any class that wants
logging functionality. Giving it a slightly different name helps
disambiguate.

In addition, this gets rid of Resque::Logging#__log__ in favor of a
class method. Boo underscore methods!