Configure failure backends with FAILURE_BACKEND
This adds a FAILURE_BACKEND environment variable which allows you to
pick which backend is used. There may be a way to do this already but if
so, I can't figure it out.

This also makes Resque::Failure::Redis the default backend and fixes
some things which broke when using that as a backend. Otherwise this
should make all of the changes fully backwards compatible (in theory!)