Fix Resque::Failure.requeue_queue/remove_queue
This includes a bit of refactoring to fix requeue_queue/remove_queue.
The logic that was previously encapsulated in Resque::Failure has been
moved to Resque::Failure::Redis (as this is the only backend that
supports requeuing/removing).

The singletons in Resque::Failure now thunk to the backend (much like
everything else) and a more intelligent backend has been substituted.
Usage of "rand" as part of the removal process has been removed o_O
with empty string being used as a sentinel value for deleted items.

This makes the test pass when using Resque::Failure::Redis as a backend,
however it's still not ready to merge as the tests are presently written
in a way which doesn't work correctly with
Resque::Failure::RedisMultiQueue. That's the next step.